ci: explictly install Clang 14 on Linux runners
The nightly build jobs run on Ubuntu 20.04 to ensure the glibc version
is compatible with WPT.fyi runners. But this version of Ubuntu has an
older clang version (10) that is not compatible with mozjs.

This patch ensures both the nightly and regular linux jobs explicity
install Clang 14 using the KyleMayes/install-llvm-action@v2 action.

Fixes #34713.

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>